feat: add ssr and rewritting some services for it

+++ music-app/src/app/core/services/data-services/local-storage-service.ts
@@ -1,19 +1,31 @@
-import { Injectable } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
+import { inject, Injectable, PLATFORM_ID } from '@angular/core';
 
 @Injectable({
   providedIn: 'root',
 })
 export class LocalStorageService {
+  private platformId = inject(PLATFORM_ID);
+
   set(key: string, value: any) {
-    localStorage.setItem(key, JSON.stringify(value));
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.setItem(key, JSON.stringify(value));
+    }
+    return null;
   }
 
   get(key: string) {
-    const item = localStorage.getItem(key);
-    return item ? JSON.parse(item) : null;
+    if (isPlatformBrowser(this.platformId)) {
+      const item = localStorage.getItem(key);
+      return item ? JSON.parse(item) : null;
+    }
+    return null;
   }
 
   clear() {
-    localStorage.clear();
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.clear();
+    }
+    return null;
   }
 }
